Refactor class transformations
Reflection is now defined before native decorators

Legacy Flamework decorators remain under the class to avoid changing execution order
